Added index.html, index2.html, stylesheet.css.
Added all html and stylesheet files for first commit. ignore index2.html, this is a revision of the original and is not complete. I plan on starting a new with a hopefully streamline or cleaner looking html. All created with the working knowledge I've acquired with FCC. So no bootstrap or anything of the like.